classroom_transcripts/pyproject.toml
[project]
name = "classroom-transcripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.37.0",
    "streamlit>=1.42.0",
    "azure-storage-blob>=12.24.1",
    "azure-identity>=1.20.0",
    "authlib>=1.4.1",
    "azure-data-tables>=12.6.0",
    "pandas>=2.2.3",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "python-docx>=1.1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "black>=25.1.0",
    "ruff>=0.9.6",
    "watchdog>=6.0.0",
    "pytest-asyncio>=0.25.3",
]

